MAJOR refactoring of the VM boot code.
